Migrated context object creation
